+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "git-hooks" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+### [1.3.1] -- 2023-08-03
+- Added support for relative path for `GitHooks.hooksDirectory` configuration
+- security fix for `GitHooks.hooksDirectory` configuration to prevent shell injection
+
 ### [1.3.0] -- 2023-07-29
 - Introduced `GitHooks` output channel to display the output logs of the extension
 - avoid await inside `activate` function of `extension.ts` to improve startup time